Consolidate .netcoreapp.cs files because System.Data.* projects is no longer cross-compiled
